Fix CS violations, another attempt to squash a connection failure occurring on the PHP7.0 test only
