+++ test/OfflineTest.php
@@ -14,6 +14,7 @@ use Zend\Config;
 use Zend\Ldap;
 use Zend\Ldap\Exception;
 use phpmock\phpunit\PHPMock;
+use ZendTest\Ldap\TestAsset\BuiltinFunctionMocks;
 
 /**
  * @group      Zend_Ldap
@@ -49,9 +50,9 @@ class OfflineTest extends \PHPUnit_Framework_TestCase
      */
     protected function activateBindableOfflineMocks()
     {
-        \LdapReusableMocks::$ldap_connect_mock->enable();
-        \LdapReusableMocks::$ldap_bind_mock->enable();
-        \LdapReusableMocks::$ldap_set_option_mock->enable();
+        BuiltinFunctionMocks::$ldap_connect_mock->enable();
+        BuiltinFunctionMocks::$ldap_bind_mock->enable();
+        BuiltinFunctionMocks::$ldap_set_option_mock->enable();
     }
 
     public function tearDown()
